% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectname-package.R
\docType{package}
\name{projectname-package}
\alias{projectname}
\alias{projectname-package}
\title{projectname: My Reproducible Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/avallecam/projectname}
  \item Report bugs at \url{https://github.com/avallecam/projectname/issues}
}

}
\author{
\strong{Maintainer}: Andree Valle-Campos \email{avallecam@gmail.com} (\href{https://orcid.org/0000-0002-7779-481X}{ORCID}) [copyright holder]

}
\keyword{internal}
